fix: renamed and fixed namespace issue for room at API proj
